backend/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.3
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
phonenumbers==8.13.33
pillow==10.2.0
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
tzdata==2024.1
